+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Added support for prime/upstream kubernetes artifacts using the `config/artifacts.yaml` file
 * Added mounting for `/usr/local` for Operating System file handling
 * Added ability to generate independent combustion drive
+* Added warning about providing container images with index digests
 
 ## API